feat: summarize chat history + organize folders (#52) (#7) (#54)
* feat: added chat history summary tool

* refactor: organized agents file

* build: added file

* refactor: initialized folder

* feat: created separate workflow file with custom streaming function

* feat: initialized + added relevant nodes

* feat: separated retrievers

* feat: separate files for model calls
