+++ src/metadata_chatbot/GAMER/_init_.py
@@ -1,0 +1,3 @@
+"""Init package"""
+
+__version__ = "0.2.2"
