--- src/metadata_chatbot/agents/__init__.py
@@ -1,3 +1,0 @@
-"""Init package"""
-
-__version__ = "0.0.84"
